// src/util/interfaces/ParsedObjects/ParsedMember.ts
export interface ParsedMember {
  day1Rank: number;
  day2Rank: number;
  day3Rank: number;
  day4Rank: number;
  day5Rank: number;
  day6Rank: number;
  day7Rank: number;
  weeklyRank: number;

  day1XP: string;
  day2XP: string;
  day3XP: string;
  day4XP: string;
  day5XP: string;
  day6XP: string;
  day7XP: string;
  weeklyXP: string;

  rank: string;
  joined: string;
}